requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Appium-Python-Client==5.0.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.3.4
redis-client==1.0.1
Redis-Sentinel-Url==1.0.1
rich==14.0.0
rich-toolkit==0.14.1
rq==1.10.1
rq-dashboard==0.8.2.2
selenium==4.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
